Llevo 9 horas aca

+++ src/Components/Gestion.js
@@ -2,6 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { Container, Row, Col, Form, Button, Table } from 'react-bootstrap';
 import axios from 'axios';
 import styles from '../Assets/gestion_empleados.module.css';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faEdit, faTrashAlt} from '@fortawesome/free-solid-svg-icons';
+import Swal from 'sweetalert2';
 
 const Gestion = () => {
   const [users, setUsers] = useState([]);
@@ -16,7 +19,6 @@ const Gestion = () => {
   const [editUser, setEditUser] = useState(null);
   const [showTable, setShowTable] = useState(false);
   const [roles, setRoles] = useState([]);
-  const [showAll, setShowAll] = useState(false);
 
   const backendUrl = 'https://proyecto2-production-ba5b.up.railway.app';
 
@@ -43,19 +45,15 @@ const Gestion = () => {
       if (response.status === 200) {
         setUsers(response.data);
         setShowTable(true);
-        setShowAll(true);
       } else {
         console.error('Error inesperado al obtener los usuarios:', response);
       }
     } catch (error) {
       if (error.response) {
-        // El servidor respondió con un código de estado fuera del rango 2xx
         console.error('Error al obtener los usuarios:', error.response.data);
       } else if (error.request) {
-        // La solicitud fue hecha pero no se recibió ninguna respuesta
         console.error('No se recibió respuesta del servidor:', error.request);
       } else {
-        // Algo sucedió en la configuración de la solicitud que desencadenó un error
         console.error('Error al configurar la solicitud:', error.message);
       }
     }
@@ -70,6 +68,7 @@ const Gestion = () => {
         }
       });
       setRoles(response.data);
+      
     } catch (error) {
       console.error('Error al obtener los roles:', error);
     }
@@ -90,6 +89,7 @@ const Gestion = () => {
           }
         });
         setEditUser(null);
+       
       } else {
         await axios.post(`${backendUrl}/api/users`, newUser, {
           headers: {
@@ -98,9 +98,19 @@ const Gestion = () => {
         });
       }
       setNewUser({ id: '', apellidos: '', nombres: '', contra: '', idRol: '' });
-      fetchUsers();
+      Swal.fire({
+        title: "¡Usuario creado!",
+        text: "El usario se creo correctamente.",
+        icon: "success"
+      });
+      fetchUsers(); // Fetch all users again to update the list
     } catch (error) {
       console.error('Error al registrar el usuario:', error);
+      Swal.fire({
+        title: "Error",
+        text: "Rol no autorizado",
+        icon: "error"
+      });
     }
   };
 
@@ -118,16 +128,41 @@ const Gestion = () => {
   const handleDelete = async (id) => {
     const token = localStorage.getItem('token'); // Obtener el token del almacenamiento local
     try {
-      await axios.delete(`${backendUrl}/api/users/${id}`, {
-        headers: {
-          'Authorization': `Bearer ${token}`
-        }
+      const result = await Swal.fire({
+        title: "Estas seguro?",
+        
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si , estoy seguro"
       });
-      fetchUsers();
+  
+      if (result.isConfirmed) {
+        await axios.delete(`${backendUrl}/api/users/${id}`, {
+          headers: {
+            'Authorization': `Bearer ${token}`
+          }
+        });
+  
+        Swal.fire({
+          title: "Usuario Eliminado",
+          
+          icon: "success"
+        });
+  
+        fetchUsers(); // Fetch all users again to update the list
+      }
     } catch (error) {
       console.error('Error al eliminar el usuario:', error);
+      Swal.fire({
+        title: "Error",
+        text: "Rol no autorizado",
+        icon: "error"
+      });
     }
   };
+  
 
   const handleSearch = async (e) => {
     e.preventDefault();
@@ -138,14 +173,24 @@ const Gestion = () => {
           'Authorization': `Bearer ${token}`
         }
       });
-      setUsers([response.data]);
+      setUsers(response.data ? [response.data] : []);
       setShowTable(true);
-      setShowAll(false);
     } catch (error) {
       console.error('Error al buscar el usuario:', error);
+      setUsers([]);
+      setShowTable(true);
     }
   };
 
+  const handleCancelEdit = () => {
+    setEditUser(null);
+    setNewUser({ id: '', apellidos: '', nombres: '', contra: '', idRol: '' });
+  };
+
+  const handleCancelSearch = () => {
+    setSearchTerm('');
+    fetchUsers();
+  };
 
   return (
     <div className={styles.gestionContainer}>
@@ -153,7 +198,7 @@ const Gestion = () => {
         <Row className="mb-5">
           <Col xs={12}>
             <h3>{editUser ? 'Editar Usuario' : 'Registrar Nuevo Usuario'}</h3>
-            <Form onSubmit={handleSubmit}>
+            <Form >
               <Form.Group>
                 <Form.Label htmlFor="id">ID</Form.Label>
                 <Form.Control type="text" id="id" name="id" value={newUser.id} onChange={handleChange} placeholder="Ingrese el ID del usuario" />
@@ -172,14 +217,15 @@ const Gestion = () => {
               </Form.Group>
               <Form.Group>
                 <Form.Label htmlFor="idRol">Rol</Form.Label>
-                <Form.Control as="select" id="idRol" name="idRol" value={newUser.idRol} onChange={handleChange} className="mb-3">
+                <Form.Control as="select" id="idRol" name="idRol" value={newUser.idRol} onChange={handleChange} className="mb-3 w-90">
                   <option value="">Seleccione un rol</option>
                   {roles.map(role => (
                     <option key={role.ID} value={role.ID}>{role.NOMBRE}</option>
                   ))}
                 </Form.Control>
               </Form.Group>
-              <Button type="submit" variant="primary" className="mb-3">{editUser ? 'Actualizar' : 'Registrar'}</Button>
+              <Button type="submit" variant="primary" className="mb-3"onClick={handleSubmit}>{editUser ? 'Actualizar' : 'Registrar'}</Button>
+              {editUser && <Button type="submit" variant="secondary" onClick={handleCancelEdit}>Cancelar</Button>}
             </Form>
           </Col>
         </Row>
@@ -187,9 +233,9 @@ const Gestion = () => {
         <Row className="mb-5">
           <Col xs={12}>
             <h3>Buscar Usuarios</h3>
-            <Form onSubmit={handleSearch}>
+            <Form>
               <Row className="mb-3">
-                <Col>
+                <Col className="w-50 ">
                   <Form.Control
                     type="text"
                     placeholder="Ingrese el ID del usuario"
@@ -197,15 +243,18 @@ const Gestion = () => {
                     onChange={(e) => setSearchTerm(e.target.value)}
                   />
                 </Col>
+              
               </Row>
-              <Row>
-                <Col xs={6} sm={3}>
-                  <Button type="submit" variant="primary" className="mb-3 w-100">Buscar</Button>
+              <Row >
+                  <Col xs="auto">
+                  <Button size="lg"  variant="primary" style={{ width: '150px' }} onClick={handleSearch}>Buscar</Button>
                 </Col>
-                <Col xs={6} sm={3}>
-                  <Button variant="primary" className="mb-3 w-100" onClick={fetchUsers}>Mostrar Todos</Button>
+                <Col xs="auto" >
+                <Button  size="lg" variant="dark" style={{ width: '150px' }} onClick={handleCancelSearch}>Cancelar</Button>
                 </Col>
+
               </Row>
+            
             </Form>
           </Col>
         </Row>
@@ -215,7 +264,7 @@ const Gestion = () => {
             <Col xs={12}>
               <h3>Lista de Usuarios</h3>
               <div className="table-responsive">
-                <Table bordered>
+                <Table bordered striped>
                   <thead className="thead-light">
                     <tr>
                       <th>ID</th>
@@ -226,31 +275,37 @@ const Gestion = () => {
                     </tr>
                   </thead>
                   <tbody>
-                    {showAll
-                      ? users.map(user => (
-                          <tr key={user.ID}>
-                            <td>{user.ID}</td>
-                            <td>{user.APELLIDOS}</td>
-                            <td>{user.NOMBRES}</td>
-                            <td>{roles.find(role => role.ID === user.IDROL)?.NOMBRE}</td>
-                            <td>
-                              <Button variant="warning" size="sm" onClick={() => handleEdit(user)}>Editar</Button>
-                              <Button variant="danger" size="sm" onClick={() => handleDelete(user.ID)}>Eliminar</Button>
-                            </td>
-                          </tr>
-                        ))
-                      : users.length > 0 && (
-                          <tr key={users[0].ID}>
-                            <td>{users[0].ID}</td>
-                            <td>{users[0].APELLIDOS}</td>
-                            <td>{users[0].NOMBRES}</td>
-                            <td>{roles.find(role => role.ID === users[0].IDROL)?.NOMBRE}</td>
-                            <td>
-                              <Button variant="warning" size="sm" onClick={() => handleEdit(users[0])}>Editar</Button>
-                              <Button variant="danger" size="sm" onClick={() => handleDelete(users[0].ID)}>Eliminar</Button>
-                            </td>
-                          </tr>
-                        )}
+                    {users.length > 0 ? (
+                      users.map(user => (
+                        <tr key={user.ID}>
+                          <td>{user.ID}</td>
+                          <td>{user.APELLIDOS}</td>
+                          <td>{user.NOMBRES}</td>
+                          <td>{roles.find(role => role.ID === user.IDROL)?.NOMBRE}</td>
+                          <td>
+                    <Button
+                      variant="warning"
+                      size="sm"
+                      onClick={() => handleEdit(user)}
+                      className="me-2"
+                    >
+                      <FontAwesomeIcon icon={faEdit} /> Editar
+                    </Button>
+                    <Button
+                      variant="danger"
+                      size="sm"
+                      onClick={() => handleDelete(user.ID)}
+                    >
+                      <FontAwesomeIcon icon={faTrashAlt} /> Eliminar
+                    </Button>
+                  </td>
+                        </tr>
+                      ))
+                    ) : (
+                      <tr>
+                        <td colSpan="5" className="text-center">Usuario Inexistente</td>
+                      </tr>
+                    )}
                   </tbody>
                 </Table>
               </div>
